fixing postgres error
