--- drizzle/index.ts
@@ -1,2 +1,0 @@
-export { db } from "./client";
-export * from "./schemas/auth";
