+++ src/core/components/user/schemas/user.ts
@@ -1,4 +1,4 @@
-import { usersTable } from "@@/drizzle";
+import { usersTable } from "@@/drizzle/schemas/auth";
 import { createInsertSchema, createSelectSchema } from "drizzle-zod";
 import { z } from "zod";
 
@@ -18,3 +18,22 @@ export const createUserSchema = createInsertSchema(usersTable).omit({
 export type CreateUser = z.infer<typeof createUserSchema>;
 
 //auth schemas
+
+export const logInSchema = z.object({
+    userIdentificator: z
+        .string()
+        .min(2, {
+            message: "El identificador debe tener al menos 2 caracteres",
+        })
+        .max(50, {
+            message: "El identificador debe tener menos de 50 caracteres",
+        }),
+    password: z
+        .string()
+        .min(8, {
+            message: "La contraseña debe tener al menos 8 caracteres",
+        })
+        .max(50, {
+            message: "La contraseña debe tener menos de 50 caracteres",
+        }),
+});
