+++ src/core/di/modules/auth-module.ts
